+++ packages/pay-common/src/api/payment.ts
@@ -61,6 +61,7 @@ export async function createPayment(
     webhookSecret,
     receiverMemo,
     apiVersion,
+    intent,
   } = params;
 
   // Set API version if provided
@@ -134,6 +135,7 @@ export async function createPayment(
     },
     ...(webhookUrl ? { webhookUrl } : {}),
     ...(webhookSecret ? { webhookSecret } : {}),
+    ...(intent ? { intent } : {}),
   };
 
   if (apiVersion === "v1") {